LinearSeparator implemented
